Increment version that works with threads
The previous try using 'register int counter asm ("r14")' does NOT simply
work with threads. On a second thought it becomes clear that
the value of r14 won't the same across each thread since the CPU will
change with context switch.

Our approach to use 'register' was precisely to avoid the problem
of, after calling the increment function, taking care of the updated
in memory value of counter inside the inline asm. But this won't work
on threads.

This new approach employs a pointer to the counter and uses its immutable
value to access the incremented value of counter in order to use it on
the comparison - cmpwi. This way the counter gets updated by the increment
function and when it returns to inline asm we refresh its value in r15 and
proceed to do the compare.

Besides that our previous increment version did not take care of LR - the
Link Register - when the inline asm calls the incremente function with 'bl',
overriding the LR value. Now we save LR value and restore it before return
from the thread main routine.