Preparation for increment in threads with HTM
Add variable 'state' used to keep track of inconsistent
behavior when non-atomic increment and state set occurs
since there is no futex, semaphore, HTM or any kind of
lock mechanism that ensure that just a single thread by
time tries to update 'counter' and 'state'.

Nonetheless, because we now perform additinal operations in
increment_counter() - i.e. update the state array, the
value of r9, being a volatile register, is lost and so
we are in trouble with:

lwa r15, 0(%0) // where %0 == r9 == counter_ptr

This issue will be fixed in the next commit.