Add a 5th example on illegal/trap instruction oddness
When a worker executes an illegal instruction the invoked signal handler
catches the illegal instruction again after returning, entering in an
infinite loop. The exactly same behavior occurs with the specific PowerPC
trap instruction.

Getting the context where the illegal/trap instruction was first executed
and incrementing the pc (program counter) seems to be a sound solution, as
discussed here on StackOverflow:

In a SIGILL handler, how can I skip the offending instruction?
http://goo.gl/7kIY1o