Add comment on IBM bit ordering to avoid insanity
POWER ISA employs IBM bit ordering, aka MSB 0, aka "Network Byte Ordering".
That must be understood, regardless of endianess - since shift (and mask)
operation is perform as per register and not memory - as a need to do a
simple converstion from MSB 0 format to LSB 0 in order avoid confusion on
shift and mask operations. Please refer to this commit content for an
example on MSR (Machine State Register) TS bits.